Split out utility functions from jqfileupload widget tests and simplified tests a bit
